requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
python-jose==3.3.0
bcrypt==4.0.1
pytest==7.4.3
pytest-cov==4.1.0
requests==2.32.3
